Run prettier; file rename
