+++ public/index.html
@@ -1,29 +1,29 @@
 <!DOCTYPE html>
-<html lang='en'>
-<style>
+<html lang="en">
+  <style>
     body {
-        background-color: black;
+      background-color: black;
     }
-</style>
-<head>
-  <meta charset='utf-8' />
-  <link
-    rel='icon'
-    href='%PUBLIC_URL%/favicon.png'
-  />
-  <link
-    rel='apple-touch-icon'
-    href='%PUBLIC_URL%/favicon.png'
-  />
-  <!--
+  </style>
+  <head>
+    <meta charset="utf-8" />
+    <link
+      rel="icon"
+      href="%PUBLIC_URL%/favicon.png"
+    />
+    <link
+      rel="apple-touch-icon"
+      href="%PUBLIC_URL%/favicon.png"
+    />
+    <!--
     manifest.json provides metadata used when your web app is installed on a
     user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
   -->
-  <link
-    rel='manifest'
-    href='%PUBLIC_URL%/manifest.json'
-  />
-  <!--
+    <link
+      rel="manifest"
+      href="%PUBLIC_URL%/manifest.json"
+    />
+    <!--
     Notice the use of %PUBLIC_URL% in the tags above.
     It will be replaced with the URL of the `public` folder during the build.
     Only files inside the `public` folder can be referenced from the HTML.
@@ -32,68 +32,71 @@
     work correctly both with client-side routing and a non-root public URL.
     Learn how to configure a non-root public URL by running `npm run build`.
   -->
-  <meta
-    name='viewport'
-    content='width=device-width, initial-scale=1'
-  />
-  <meta
-    name='theme-color'
-    content='#000000'
-  />
-  <meta
-    name='version'
-    content='%REACT_APP_GIT_HASH%'
-  />
-  <meta
-    name='description'
-    content='Sarcophagus is an autonomous, incentivized application built on Ethereum and Arweave.'
-  />
-  <meta
-    property='og:url'
-    content='%PUBLIC_URL%'
-  />
-  <meta
-    property='og:description'
-    content='Sarcophagus is an autonomous, incentivized application built on Ethereum and Arweave.'
-  />
-  <meta
-    property='og:type'
-    content='website'
-  />
-  <meta
-    property='og:title'
-    content='Sarcophagus'
-  />
-  <meta
-    property='og:image'
-    content='%PUBLIC_URL%/favicon.png'
-  />
-  <meta
-    property='og:image:alt'
-    content='Sarcophagus'
-  />
-  <meta
-    property='og:site_name'
-    content='Sarcophagus'
-  />
-  <title>Sarcophagus</title>
-  <script async src='https://www.googletagmanager.com/gtag/js?id=G-EVVFRT1G00'></script>
-  <script>
-    window.dataLayer = window.dataLayer || [];
+    <meta
+      name="viewport"
+      content="width=device-width, initial-scale=1"
+    />
+    <meta
+      name="theme-color"
+      content="#000000"
+    />
+    <meta
+      name="version"
+      content="%REACT_APP_GIT_HASH%"
+    />
+    <meta
+      name="description"
+      content="Sarcophagus is an autonomous, incentivized application built on Ethereum and Arweave."
+    />
+    <meta
+      property="og:url"
+      content="%PUBLIC_URL%"
+    />
+    <meta
+      property="og:description"
+      content="Sarcophagus is an autonomous, incentivized application built on Ethereum and Arweave."
+    />
+    <meta
+      property="og:type"
+      content="website"
+    />
+    <meta
+      property="og:title"
+      content="Sarcophagus"
+    />
+    <meta
+      property="og:image"
+      content="%PUBLIC_URL%/favicon.png"
+    />
+    <meta
+      property="og:image:alt"
+      content="Sarcophagus"
+    />
+    <meta
+      property="og:site_name"
+      content="Sarcophagus"
+    />
+    <title>Sarcophagus</title>
+    <script
+      async
+      src="https://www.googletagmanager.com/gtag/js?id=G-EVVFRT1G00"
+    ></script>
+    <script>
+      window.dataLayer = window.dataLayer || [];
 
-    function gtag() {
-      dataLayer.push(arguments);
-    }
+      function gtag() {
+        dataLayer.push(arguments);
+      }
 
-    gtag('js', new Date());
+      gtag('js', new Date());
 
-    gtag('config', 'G-EVVFRT1G00');
-  </script>
-</head>
-<body>
-<noscript>You need to enable JavaScript to run this app.</noscript>
-<div id='root'></div>
-<!--
+      gtag('config', 'G-EVVFRT1G00');
+    </script>
+  </head>
+  <body>
+    <noscript>You need to enable JavaScript to run this app.</noscript>
+    <div id="root"></div>
+    <!--
   This HTML file is a template.
   If you open it directly in the browser, you will see an empty page.
 
@@ -103,5 +106,5 @@
   To begin the development, run `npm start` or `yarn start`.
   To create a production bundle, use `npm run build` or `yarn build`.
 -->
-</body>
+  </body>
 </html>
